refactor: Move approval models to schemas module
