+++ tests/unit/test_approvals_service.py
@@ -5,7 +5,7 @@ import uuid
 import pytest
 from sqlalchemy.exc import IntegrityError
 from sqlmodel.ext.asyncio.session import AsyncSession
-from tracecat_ee.agent.approvals.models import ApprovalCreate, ApprovalUpdate
+from tracecat_ee.agent.approvals.schemas import ApprovalCreate, ApprovalUpdate
 from tracecat_ee.agent.approvals.service import ApprovalService, SessionInfo
 
 from tracecat.agent.approvals.enums import ApprovalStatus
